WIP Add La Crosse LTV omnibus decoder
The protocols for all the reverse-engineered 915 MHz La Crosse
Technology LTV sensors share a lot of commonality. They all follow the
same general format with the same modulation, symbol rate,
synchronization pattern, device id field and CRC polynomial. The
existing decoders have resorted to heuristics such as measuring the
length of the preamble and trailer to infer which model transmitted the
message. These heuristics are far from perfect and suffer from false
positives where a decoder parses packets from one model as if they were
from an incompatible model, producing garbage output alongside the
correct decoder.

Given that the only identifying information in the packets is the device
id, La Crosse's devices and cloud services must be determining the model
that transmitted it, and thus how to decode the rest of the packet, from
the device id. There is speculation that the model is encoded in the
most-significant bits of the device id. The corpus of known device ids
is consistent with this theory, though more data is needed to confirm.

Add a decoder for all La Crosse LTV sensors which uses best guesses
about the mapping of device id to model to parse the packet contents.